refactor: extract listening functionality into Stop handler
